server/requirements.txt
Flask==3.1.2
Pillow==12.0.0
flask-cors==5.0.0
numpy
numba
opencv-python-headless
scipy
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
